chore: dev 브랜치 최신 사항 반영

+++ src/main/java/com/example/palayo/domain/deposithistory/dto/DepositHistoryResponse.java
@@ -14,7 +14,7 @@ public class DepositHistoryResponse {
     private Long id;
     private Long auctionId;
     private Long userId;
-    private Long deposit;
+    private int deposit;
     private DepositStatus status;
     private LocalDateTime createdAt;
